Unify Chart and Candlestick Patterns into Trading Academy
- Consolidate modals into `AcademyModal.svelte`
- Refactor `CandlestickPatternsModal` -> `CandlestickPatternsView`
- Refactor `ChartPatternsModal` -> `ChartPatternsView`
- Update `uiState` to replace separate flags with `showAcademyModal`
- Update `LeftControlPanel` to use a single button
- Enhance `ChartPatternChart` visuals with grid/axes
- Update localization with "Trading Academy" title
